chore: Refactor db and dependencies setup to make sure the same connection pool is used when creating new connections
