+++ app/commands/check_expired_invitations.py
@@ -1,5 +1,4 @@
 import asyncio
-from contextlib import asynccontextmanager
 from datetime import UTC, datetime
 
 import typer
@@ -7,16 +6,13 @@ from rich import print
 from sqlalchemy import update
 
 from app.conf import Settings
-from app.db.base import get_db_engine, get_db_session, get_db_sessionmaker
+from app.db.base import session_factory
 from app.db.models import AccountUser
 from app.enums import AccountUserStatus
 
 
 async def check_expired_invitations(settings: Settings):
-    engine = get_db_engine(settings)
-    session_maker = get_db_sessionmaker(engine)
-
-    async with asynccontextmanager(get_db_session)(session_maker) as session:
+    async with session_factory.begin() as session:
         stmt = (
             update(AccountUser)
             .where(
