+++ app/commands/invite_user.py
@@ -1,6 +1,5 @@
 import asyncio
 import secrets
-from contextlib import asynccontextmanager
 from datetime import UTC, datetime, timedelta
 from typing import Annotated
 
@@ -9,7 +8,7 @@ from email_validator import EmailNotValidError, validate_email
 from rich import print
 
 from app.conf import Settings
-from app.db.base import get_db_engine, get_db_session, get_db_sessionmaker
+from app.db.base import session_factory
 from app.db.handlers import AccountHandler, AccountUserHandler, UserHandler
 from app.db.models import Account, AccountUser, User
 from app.enums import AccountStatus, AccountType, AccountUserStatus, UserStatus
@@ -66,10 +65,7 @@ async def invite_user(
     settings: Settings, email: str, name: str, account_id: str | None, force: bool = False
 ):
     """ """
-    engine = get_db_engine(settings)
-    session_maker = get_db_sessionmaker(engine)
-
-    async with asynccontextmanager(get_db_session)(session_maker) as session:
+    async with session_factory.begin() as session:
         account_handler = AccountHandler(session)
         user_handler = UserHandler(session)
         accountuser_handler = AccountUserHandler(session)
