+++ app/conf.py
@@ -1,8 +1,6 @@
 import pathlib
 from functools import lru_cache
-from typing import Annotated
 
-from fastapi import Depends
 from pydantic import PostgresDsn, computed_field
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
@@ -71,6 +69,3 @@ class Settings(BaseSettings):
 @lru_cache
 def get_settings() -> Settings:
     return Settings()
-
-
-AppSettings = Annotated[Settings, Depends(get_settings)]
