+++ tests/commands/test_invite_user.py
@@ -17,14 +17,6 @@ from app.enums import AccountStatus, AccountUserStatus, UserStatus
 from tests.types import ModelFactory
 
 
-@pytest.fixture(autouse=True)
-def mock_db_session(db_session: AsyncSession, mocker: MockerFixture):
-    async def mock_get_db_session(*args, **kwargs):  # noqa: RUF029
-        yield db_session
-
-    mocker.patch("app.commands.invite_user.get_db_session", new=mock_get_db_session)
-
-
 @time_machine.travel("2025-03-07T10:00:00Z", tick=False)
 async def test_invite_user(
     test_settings: Settings,
